added unit tests for echo request handler
#What: added unit tests for a successful and unsuccessful requests to echo.

#Why: adding unit tests to ensure code functionality.

#How: created a test fixture and new testing files for the echo request handler. works similarly to how our other tests are configured.

#Test: test suite passes with these added!

+++ include/echo_request_handler.h
@@ -7,7 +7,7 @@
 class Echo_Request_Handler{
     public:
         Echo_Request_Handler();
-        reply handleRequest(char* data, size_t bytes_transferred, reply::status_type status);
+        reply handleRequest(char* data, int bytes_transferred, reply::status_type status);
     private:
         reply reply_;
 };
